Fixed null return & session close

+++ src/main/java/internetshop/controllers/DeleteItemFromBucketController.java
@@ -1,8 +1,10 @@
 package internetshop.controllers;
 
 import internetshop.annotations.Inject;
+import internetshop.model.Bucket;
 import internetshop.service.BucketService;
 import java.io.IOException;
+import java.util.Optional;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -16,8 +18,11 @@ public class DeleteItemFromBucketController extends HttpServlet {
             throws IOException {
         Long userId = (Long) request.getSession(true).getAttribute("userId");
         Long itemToDeleteId = Long.valueOf(request.getParameter("item_id"));
-        Long bucketId = bucketService.getByUserId(userId).getBucketId();
-        bucketService.deleteItem(bucketId, itemToDeleteId);
+        Optional<Bucket> bucket = bucketService.getByUserId(userId);
+        if (bucket.isPresent()) {
+            Long bucketId = bucket.get().getBucketId();
+            bucketService.deleteItem(bucketId, itemToDeleteId);
+        }
 
         response.sendRedirect(request.getContextPath() + "/user/show-bucket");
     }
